test!: separate channel and unit tests (#211)
* ci!: tox -e py runs unit tests only
  (lowers local test coverage!)
* refactor: simplify STM construction in tests
* refactor: remove unnecessary Test classes
* test: remove f2 states from helicity recipe test
  (speeds up unit tests by 1 second)
* test: reorganise in order to follow module structure
* fix: update expected_recipe.yml
