+++ doc/conf.py
@@ -76,7 +76,6 @@ exclude_patterns = [
     "**.ipynb_checkpoints",
     "*build",
     "adr/template.md",
-    "test",
     "tests",
 ]
