+++ tests/channels/__init__.py
@@ -1,1 +1,1 @@
-"""Required to ignore mypy errors in the tests folder."""
+"""Required to ignore mypy errors in the :file:`tests` folder."""
