+++ tests/unit/io/test_particle_collection.py
