+++ tests/unit/io/test_xml_adapters.py
