+++ tests/unit/state/conservation_rules/test_ls_coupling.py
