+++ tests/unit/state/conservation_rules/test_spin.py
