+++ tests/unit/topology/test_simple_state_builder.py
