+++ tests/unit/ui/test_system_control.py
