+++ tox.ini
@@ -16,9 +16,8 @@ envlist =
 whitelist_externals =
     pytest
 commands =
-    pytest tests {posargs} \
-        -m "not slow" \
-        --cov-fail-under=85 \
+    pytest {posargs:tests/unit} \
+        --cov-fail-under=69 \
         --cov-report=html \
         --cov-report=xml \
         --cov=expertsystem
@@ -62,7 +61,7 @@ commands =
 whitelist_externals =
     pytest
 commands =
-    pytest tests {posargs} \
+    pytest {posargs:tests} \
         --cov-fail-under=85 \
         --cov-report=html \
         --cov-report=xml \
